requirements.txt
# Azure OpenAI
openai>=1.12.0

# Azure AI Search
azure-search-documents>=11.4.0
azure-core>=1.29.0

# Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Utilities
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
